% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq_taxa.R
\name{load_seq_taxa}
\alias{load_seq_taxa}
\title{Load Sequence and Taxa Tables}
\usage{
load_seq_taxa(date = "230212")
}
\arguments{
\item{date}{Character value of the date associated with dataset.}
}
\value{
Files are loaded into environment
}
\description{
Load Sequence and Taxa Tables
}
